Add filters module and refactor main.py for improved chat functionality and database integration
